requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4 == 4.8.2
lxml == 4.5.0
python_telegram_bot == 20.0a4
requests == 2.22.0
